refactor(test): rename client and client methods
rename client and client methods
