+++ packages/integration-tests/tests/resource.test.ts
